refactor: split client.go into per-domain files
client.go had grown to 1107 lines, but only 46 of its declarations were
real: 20 Client methods, 18 functions and 8 types. The other 133 were
pass-through re-exports. Its own section banners already marked the
seams — this turns them into files.

  client.go     237  Client, NewClient, options, the store chain
  repo.go       304  init, format upgrade, key slots, password/recovery
  aliases.go     76  the re-exported types
  backup.go     193  backup, profiles, workstation setup
  query.go      165  list, ls, find, diff, cat
  retention.go  117  prune, forget, break-lock
  restore.go     47
  check.go       30

Purely structural: `go doc -all` reports the same 119 exported symbols
before and after, byte for byte, which is the check that this changed
where code lives and nothing else.

The option-completeness guard added earlier read client.go by name and
failed the moment the re-exports moved — exactly what it is for. It now
scans every file in the root package, so it cannot be defeated by moving
a var block to a new file.
